[変更]インスタンス管理を変更
